requirements.txt
asgiref==3.5.1
Django==4.0.4
django-ckeditor==6.4.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
